refactor: clean up textrecognizer
